+++ include/render/denoise/BilateralDenoiser.h
@@ -10,7 +10,7 @@ namespace render {
     std::unique_ptr<Denoiser> clone() const override;
     const char* diagnosticName() const override;
     DenoiserDiagnostics diagnostics() const override;
-    void denoise(Buffer<Colord>& buffer) const override;
+    void denoiseFrame(DenoiserFrame& frame) const override;
 
     void setRadius(int radius);
     int radius() const;